### 2019-09-13 #1728 @whedon
[PRE REVIEW]: GKNAP: A Java and C++ package for solving the multidimensional knapsack problem
**Submitting author:** @shah314 (<a href="http://orcid.org/0000-0002-3770-1391">Shalin S Shah</a>)
**Repository:** <a href="https://github.com/shah314/gamultiknapsack" target ="_blank">https://github.com/shah314/gamultiknapsack</a>
**Version:** v1.6
