### 2019-09-13 1e2c4d8
Update paper.md
