### 2019-09-13 f5c6b70
Update paper.md
