+++ tests/README.md
@@ -1,0 +1,2 @@
+<h1>GNAP Tests</h1>
+Please run the test.sh script after cloning the repository. It will test the Java as well as the C++ code.
