+++ tests/README.md
@@ -1,5 +1,6 @@
 <h1>GNAP Tests</h1>
 Please run the bash scripts after cloning the repository. You may have to change the bash binary path in the scripts.
+<br>
 <ul>
   <li>Run testjava.sh to test the java code</li>
   <li>Run testcpp.sh to test the C++ code</li>
