### 2019-09-21 #1728 @arfon
> @stulacy could also be a reviewer

:wave: @stulacy - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
