+++ tests/README.md
@@ -1,6 +1,8 @@
 <h1>GNAP Tests</h1>
 Please run the bash scripts after cloning the repository. You may have to change the bash binary path in the scripts.<br>
-<ul>
+<br><ul>
   <li>Run testjava.sh to test the java code</li>
   <li>Run testcpp.sh to test the C++ code</li>
 </ul>
+<br>
+Please note that since the code is randomized, it will not find the optimum every time it is run (unless it is run for a very large number of iterations).
