### 2019-09-25 #1756 @stulacy
Can you please add more documentation on how to use the program?

For example, be explicit in that you need to be in the `java/classes` directory to run the `java GeneticAlgorithm` command in the first place (imo executables should be in a high-level folder clearly marked, it took me a minute to find these).
Mention how to format your data into the .dat format required.
How to interpret the output (i.e. is it just printed out to standard out or can it be obtained programmatically from the java function?)
Are there any arguments to this or do you only ever need to run this exact command with a correctly formatted dataset in the same directory?

One functional change that would be very much appreciated is the ability to pass in the filename of the data file rather than having to rename the file each time you want to run a different problem.

There's no mention of testing - how do I know the program is working correctly? I assume the default dataset in the `classes` folder is `Weing1` as I get the same output as mentioned in the README but this isn't explicitly stated anywhere.

Also mention how to install it. Are you expecting users to just run the compiled Java code themselves? If so what version of Java do they need? Are there any other dependencies?
The C++ version wouldn't compile for me, a truncated log is below.

```
(base) stuart@bishy ~/repos/gamultiknapsack $ g++ cmultiknapsack.cpp
cmultiknapsack.cpp: In function ‘void processDataORLIB()’:
cmultiknapsack.cpp:691:15: error: ‘strtok’ was not declared in this scope; did you mean ‘strtoq’?
  691 |  char * tok = strtok(line, " ");
      |               ^~~~~~
      |               strtoq
cmultiknapsack.cpp:719:7: error: ‘strcmp’ was not declared in this scope
  719 |    if(strcmp(tok, " ") == 0 || strcmp(tok, "\n") == 0)
      |       ^~~~~~
cmultiknapsack.cpp:35:1: note: ‘strcmp’ is defined in header ‘<cstring>’; did you forget to ‘#include <cstring>’?
   34 | #include <time.h>
  +++ |+#include <cstring>
   35 |
cmultiknapsack.cpp:764:16: error: ‘strcmp’ was not declared in this scope
  764 |             if(strcmp(tok, " ") == 0 || strcmp(tok, "\n") == 0)
      |                ^~~~~~
cmultiknapsack.cpp:764:16: note: ‘strcmp’ is defined in header ‘<cstring>’; did you forget to ‘#include <cstring>’?
cmultiknapsack.cpp:803:13: error: ‘strcmp’ was not declared in this scope
  803 |          if(strcmp(tok, " ") == 0 || strcmp(tok, "\n") == 0)
      |             ^~~~~~
cmultiknapsack.cpp:803:13: note: ‘strcmp’ is defined in header ‘<cstring>’; did you forget to ‘#include <cstring>’?
```
