### 2019-09-25 a9452b1
Update cmultiknapsack.cpp
