### 2019-09-25 edcddd6
Update cmultiknapsack.cpp
