+++ tests/README.md
@@ -5,7 +5,7 @@ Please run the bash scripts after cloning the repository. You may have to change
   <li>Run testcpp.sh to test the C++ code</li>
 </ul>
 <br>
-Please note that since the code is randomized, it will not find the optimum every time it is run (unless it is run for a very large number of iterations). Also, the data.DAT file is the WEING1 instance.
+Please note that since the code is randomized, it will not find the optimum every time it is run (unless it is run for a very large number of iterations). The data.DAT file is the WEING1 instance. If the code gets stuck in a solution which is not the optimum for a very long time, please terminate it and re-run.
 <br><br>
 The benchmark instances are available <a href="http://people.brunel.ac.uk/~mastjjb/jeb/orlib/files/mknap2.txt">here</a>. They have the following format:
 <pre>
