+++ cmultiknapsack.cpp
@@ -3,8 +3,8 @@
 Genetic Algorithm - 0/1 Multi-Constraint Knapsack Problem (Multi-Dimensional)
 -----------------------------------------------------------------------------
 
-A genetic algorithm implementation for the multi-constraint knapsack problem. 
-The multi-constraint knapsack problem is a generalization of the 0/1 knapsack problem. The 0/1 
+A genetic algorithm implementation for the multi-constraint knapsack problem.
+The multi-constraint knapsack problem is a generalization of the 0/1 knapsack problem. The 0/1
 knapsack problem has one weight constraint. The multi-constraint knapsack problem has m constraints
 and one objective function to be maximized while all the m constraints are satisfied.
 This makes it much harder to solve as compared to the 0/1 knapsack problem.
@@ -98,17 +98,17 @@ int * VALUES;  // populated automatically by processDataORLIB
 const int POPULATION = 10; // size of the population
 const int LOCAL_IMPROVEMENT = 10; // number of local improvements
 const double INITIAL_POPULATION_PROB = 0.9; // the probability with which the initial population is generated
-const int GENERATIONS = 100000; // number of generations to run the algorithm
+const int GENERATIONS = 100; // number of generations to run the algorithm
 const int GREEDY_CROSSOVER = 1; // a constant to identify greedy crossover in adaptive crossover
 const int ONE_POINT_CROSSOVER = 0; // a constant to identify one point crossover in adaptive crossover
 const int NEGATIVE_FITNESS = -100; // the fitness of an invalid individual (violating constraints)
-const int SHUFFLE_TOLERANCE = 2; // the number of attempts to identify unique parents before shuffling
+const int SHUFFLE_TOLERANCE = 100000; // the number of attempts to identify unique parents before shuffling
 double MUTATION_PROBABILITY;	// populated in main()
 double MUTATION_INCREMENT;	// populated in main()
 double SHUFFLE_PROBABILITY;	// populated in main()
 double MAX_MUTATION_PROBABILITY;	// populated in main()
 int runtimeCrossoverType; // populated by adaptive crossover to identify which crossover operation was performed
-const int MAX_UNIQUE_ITERATIONS = 10000; // maximum iterations to spend on finding objects in local improvement
+const int MAX_UNIQUE_ITERATIONS = 10; // maximum iterations to spend on finding objects in local improvement
 
 /* A list of objects sorted in non-increasing order of the lagrangian psuedo-utility ratio.
 Used by localImprovement() */
@@ -679,10 +679,10 @@ vector<KNode> generateRandomPopulation()
 }
 
 /* Process the data from the ORLIB file */
-void processDataORLIB()
+void processDataORLIB(char * filename)
 {
 	FILE * file;
-	file = fopen("orlib1.txt", "r");
+	file = fopen(filename, "r");
    if(file == NULL)
    {
    	printf("orlib1.txt File Not Found in Current Directory.");
@@ -701,7 +701,7 @@ void processDataORLIB()
 	//printf("%d\n", NUMBER_CONSTRAINTS);
 	//printf("%d\n", OPTIMUM);
[…]
