+++ java/DataProcessorORLIB.java
@@ -7,14 +7,14 @@ import java.io.*;
  * @author Shalin Shah
  * Email: shah.shalin@gmail.com
  */
-public class DataProcessorORLIB 
+public class DataProcessorORLIB
 {
-    /** 
+    /**
      * Process the data from the file specified by the constant ORLIB_FILE in Constants.java
      */
-    public static void processData() throws Exception
+    public static void processData(String filename) throws Exception
     {
-        File file = new File(Constants.ORLIB_FILE);
+        File file = new File(filename);
         BufferedReader reader = new BufferedReader(new FileReader(file));
         String numOBJ = reader.readLine();
         StringTokenizer token = new StringTokenizer(numOBJ, " ");
@@ -42,7 +42,7 @@ public class DataProcessorORLIB
                 break;
             }
         }
-        
+
         Constants.CONSTRAINTS = new int[Constants.NUMBER_CONSTRAINTS][Constants.NUMBER_OBJECTS];
         for(i=0; i<Constants.NUMBER_CONSTRAINTS; i++)
         {
@@ -56,8 +56,8 @@ public class DataProcessorORLIB
                 {
                     Constants.CONSTRAINTS[i][j] = Integer.parseInt(token.nextToken().trim());
                     j++;
-                }   
-                
+                }
+
                 if(j < Constants.NUMBER_OBJECTS)
                 {
                     constraint = reader.readLine();
@@ -69,12 +69,12 @@ public class DataProcessorORLIB
                 }
             }
         }
-        
+
         String capacities = reader.readLine();
         token = new StringTokenizer(capacities, " ");
         Constants.CAPACITIES = new int[Constants.NUMBER_CONSTRAINTS];
         i=0;
-        
+
         while(true)
         {
             while(token.hasMoreTokens())
@@ -82,7 +82,7 @@ public class DataProcessorORLIB
                 Constants.CAPACITIES[i] = Integer.parseInt(token.nextToken().trim());
                 i++;
             }
-            
+
             if(i < Constants.NUMBER_CONSTRAINTS)
             {
                 capacities = reader.readLine();
