+++ tests/testjava.sh
@@ -4,7 +4,7 @@ cd testing1
 cp ../../java/*.java .
 cp ../data.DAT .
 javac *.java
-java GeneticAlgorithm
+java GeneticAlgorithm data.DAT weing
 cd ..
 rm testing1/*.java
 rm testing1/*.class
