### 2019-09-25 6b45f46
Update GreedyAlgorithm.java
