+++ README.md
@@ -17,21 +17,38 @@ The <b>orlib</b> formatted files are available <a href="http://people.brunel.ac.
 
 <b>Java implementation</b>
 Any JDK should work fine. If the code does not compile, please open an issue.
-Compile the Java code and then run:
-<b>java GeneticAlgorithm filename format</b>
+Compile the Java code and then run GeneticAlgorithm.
+
+```
+javac *.java
+java GeneticAlgorithm filename format
+```
+
 (The file name contains the instance in weing or orlib format)
 (The format is either weing or orlib)
-Example: <b>java GeneticAlgorithm data.DAT weing</b>
+Example: 
+
+```
+java GeneticAlgorithm data.DAT weing
+```
 
 <b>C++ implementation</b>
 The code was tested on a Mac with gcc version 8, downloaded using homebrew. 
 If the code does not compile, please open an issue.
 Compile the C++ code and then run the executable.
+
+```
 g++ cmultiknapsack.cpp
-<b>./a.out filename format</b>
+./a.out filename format
+```
+
 (The file name contains the instance in weing or orlib format)
 (The format is either weing or orlib)
-Example: <b>./a.out data.DAT weing</b>
+Example: 
+
+```
+./a.out data.DAT weing
+```
 
 (Please remove all comments and other extraneous text from data.DAT)
 (See the tests directory for <b>testcpp.sh and testjava.sh</b> for an example run)
