### 2019-09-25 8e0693e
Update testcpp.sh
