+++ tests/testcpp.sh
@@ -5,8 +5,3 @@ cp ../../*.cpp .
 cp ../data.DAT .
 g++ *.cpp
 ./a.out data.DAT weing
-cd ..
-rm testing1/*.cpp
-rm testing1/*.out
-rm testing1/data.DAT
-rmdir testing1
