### 2019-09-25 c934983
Update testjava.sh
