### 2019-09-26 #1756 @shah314
@stulacy I think I have addressed all of your suggestions. The code has much more documentation in the README.md.

I am now passing in the file name and the format for both Java and C++ versions.

The java/classes directory is only for convenience, but the user should really compile the code on their own machine. C++ should also compile now (I tested it on g++ version 8). There were some headers missing, which was ok when using the Mac C++ compiler, but you may be using some other version of g++.

I have added some test scripts in the tests directory which will be helpful in understanding how the code works. I have also described the format of the file that the code expects.

There is no installation. Just compile the code and run. There are also no dependencies.

To get the output programmatically, the user can look at the main methods in both the Java and C++ code.

The Java code was tested using JDK 1.8 and the C++ code was tested using g++ version 8. I have added this to the README.

Please do a git pull and verify and review the changes.
