### 2019-09-28 #1756 @stulacy
Thank you @shah314 , I feel the changes help make the software more user friendly.
My only remaining comment is that I cannot see anywhere where the difference between Orlib and Weing data encoding is explicitly stated. Would it be useful to have this information displayed somewhere?
