+++ README.md
@@ -18,8 +18,8 @@ Please note that these files contain multiple instances, so to run the algorithm
 please use any <b>one</b> of the instances.
 
 <b>Java implementation</b>
-I tested the code using JDK 1.8, but any JDK should work fine. If the code does not compile, 
-please open an issue. 
+I tested the code using JDK 1.8, but any JDK should work fine. 
+If the code does not compile, please open an issue. 
 
 Compile the Java code and then run GeneticAlgorithm.
 
