+++ README.md
@@ -17,6 +17,9 @@ The orlib formatted files are available <a href="http://people.brunel.ac.uk/~mas
 Please note that these files contain multiple instances, so to run the algorithm, 
 please use any <b>one</b> of the instances.
 
+The distributions of the files are quite different between orlib and weing. 
+If the algorithm is stuck, please increase Constants.DIFF_TOLERANCE on line 31 of GeneticAlgorithm.java.
+
 <b>Java implementation</b>
 I tested the code using JDK 1.8, but any JDK should work fine. 
 If the code does not compile, please open an issue. 
