+++ tests/README.md
@@ -10,14 +10,16 @@ Please note that since the code is randomized, it will not find the optimum ever
 <pre>
 The <b>weing</b> formatted files are available <a href="http://people.brunel.ac.uk/~mastjjb/jeb/orlib/files/mknap2.txt">here</a>.
 The <b>orlib</b> formatted files are available <a href="http://people.brunel.ac.uk/~mastjjb/jeb/orlib/mknapinfo.html">here</a> (used by P.C.Chu and J.E.Beasley).
+
 The orlib formatted files are available <a href="http://people.brunel.ac.uk/~mastjjb/jeb/orlib/files/">here</a> which are named mknapcb1.txt, mknapcb2.txt and so on.
+
 Please note that these files contain multiple instances, so to run the algorithm, 
 please use any <b>one</b> of the instances.
 
 The distributions of the files are quite different between orlib and weing. 
 If the algorithm is stuck, please increase Constants.DIFF_TOLERANCE on line 31 of GeneticAlgorithm.java.
 </pre>
-<br><br>
+<br>
 <pre>
  //This is the WEING1.DAT data file plus some comments, that are NOT part of the problem instance.
  
