### 2019-09-28 #1756 @shah314
Thank you. I have added links to the page where all files are described. The bottom of the page also contains a link to all files: http://people.brunel.ac.uk/~mastjjb/jeb/orlib/mknapinfo.html
