### 2019-09-29 4fea39c
Update GreedyAlgorithm with orlib constants
Update GreedyAlgorithm with orlib constants
