### 2019-10-03 #1756 @sjvrijn
I have some remarks and questions:

Community Guidelines:
- CONTRIBUTING.md mentions 'graph coloring code' instead of gamultiknapsack
- How should interested third parties seek support?

Paper
- Summary:
  - [line 2-3] "It is a strongly NP-hard problem, and [it] might be difficult to solve [it] using ..." the "it"s in square brackets can be removed.
- Mathematics:
  - [line 8] expect a colon ':' before this formula, a trailing comma
  - [line 9] 'Where' should be lowercase, missing 'and' after the comma

  - [line 12] "Our method [...] is described in (Shah, 2019)" This seems to conflict with [Summary line 14-15] "Our algorithm is *similar to* (Shah, 2019)". Is there a difference? If so, what is it?
  - [paragraph 6] The difference between cases with and without domain knowledge is described, but I cannot determine which case your algorithm is catering to. Is it domain-knowledge informed? If so, what domain knowledge is needed? I can't find this in the README either.

Repo documentation README
- [Cited by] How can your paper from 2019 be cited by two papers from 2012? What is being cited by whom exactly? I also cannot find the first of the two referenced papers.
