### 2019-10-03 ee36bab
Update CONTRIBUTING.md
