### 2019-10-03 b595f4d
Update paper.md
