### 2019-10-03 #1756 @shah314
@sjvrijn Thank you for the comments. I think I have made the changes you suggested.

About the (Shah, 2019) reference, the algorithm is similar (not the same). The cited paper solves the 0/1 knapsack problem while this paper solves the multidimensional knapsack problem. So, though similar, the two applications are entirely different. I changed the text in the paper.

About domain knowledge, what I meant was that a utility ratio is available, which might not be available in other problems.

About the citations, this code was written in a very basic form several years back. The code was cited by these papers and not the paper. I am also not able to find the first of the two papers. To make this more current, I have removed the "Cited By" section of the README.
