### 2019-10-13 shah314/gamultiknapsack#2 @sjvrijn
[java] program hangs on 1st case from mknap1 after showing greedy result
When running the Java version on the first case from the [`mknap1`](http://people.brunel.ac.uk/~mastjjb/jeb/orlib/files/mknap1.txt)-set, the program prints the greedy result, but just hangs after that.

The test cases included in the tests folder run normally.

Java output:
```
/d/src/gamultiknapsack/java$ java.exe GeneticAlgorithm ../files/mknap1-1.txt orlib
Calculating Lambdas...
Finished
Greedy Solution: 3800
<program hangs here for at least 10 minutes>
```

C++ output for comparison:
```
/d/src/gamultiknapsack$ ./a.exe files/mknap1-1.txt orlib
Data file format: orlib
Data file name: files/mknap1-1.txt
0:-100
1:3800
Optimum Found!
Solution Objective Function: 3800
Objects chosen:
1 2 5
```

Extracted test case [mknap1-1.txt](https://github.com/shah314/gamultiknapsack/files/3722464/mknap1-1.txt):
```
 6 10 3800
 100 600 1200 2400 500 2000
 8 12 13 64 22 41
 8 12 13 75 22 41
 3 6 4 18 6 4
 5 10 8 32 6 12
 5 13 8 42 6 20
 5 13 8 48 6 20
 0 0 0 0 8 0
 3 0 4 0 8 0
 3 2 4 0 8 4
 3 2 4 8 8 4
 80 96 20 36 44 48 10 18 22 24
```

OS: Windows 10
```
$ java.exe --version
openjdk 13 2019-09-17
OpenJDK Runtime Environment (build 13+33)
OpenJDK 64-Bit Server VM (build 13+33, mixed mode, sharing)
```
