### 2019-10-13 shah314/gamultiknapsack#3 @sjvrijn
[java] NumberFormatException on test-case with floating-point numbers
When running the Java version on the second case from the [mknap1](http://people.brunel.ac.uk/~mastjjb/jeb/orlib/files/mknap1.txt)-set, the program prints the following error:

```
/d/src/gamultiknapsack/java$ java.exe GeneticAlgorithm ../files/mknap1-2.txt orlib
Exception in thread "main" java.lang.NumberFormatException: For input string: "8706.1"
        at java.base/java.lang.NumberFormatException.forInputString(NumberFormatException.java:68)
        at java.base/java.lang.Integer.parseInt(Integer.java:658)
        at java.base/java.lang.Integer.parseInt(Integer.java:776)
        at DataProcessorORLIB.processData(DataProcessorORLIB.java:23)
        at GeneticAlgorithm.main(GeneticAlgorithm.java:32)
```

C++ output for comparison:
```
/d/src/gamultiknapsack$ ./a.exe files/mknap1-2.txt orlib
Data file format: orlib
Data file name: files/mknap1-2.txt
0:-100
1:8335
[...]
5:8335
6:8576
[...]
12:8576
13:8593
[...]
19:8593
20:8705
[...]
99:8705
Solution Objective Function: 8705
Objects chosen:
1 3 4 7 9
```

Extracted test case [mknap1-2.txt](https://github.com/shah314/gamultiknapsack/files/3722476/mknap1-2.txt):
```
 10 10 8706.1
 600.1 310.5 1800 3850 18.6 198.7 882 4200 402.5 327
 20 5 100 200 2 4 60 150 80 40
 20 7 130 280 2 8 110 210 100 40
 60 3 50 100 4 2 20 40 6 12
 60 8 70 200 4 6 40 70 16 20
 60 13 70 250 4 10 60 90 20 24
 60 13 70 280 4 10 70 105 22 28
 5 2 20 100 2 5 10 60 0 0
 45 14 80 180 6 10 40 100 20 0
 55 14 80 200 6 10 50 140 30 40
 65 14 80 220 6 10 50 180 30 50
 450 540 200 360 440 480 200 360 440 480
```
