### 2019-10-13 #1756 @sjvrijn
@shah314 Thanks for the updates, the paper and additional documentation files seem fine to me now.

I encountered two problems when running some test files on the Java version, for which I have opened Issues in the repository: [program hangs on simple case](https://github.com/shah314/gamultiknapsack/issues/2) and [NumberFormatException on float-input](https://github.com/shah314/gamultiknapsack/issues/3)

Additionally, some of the documentation comments around the input files in both the C++ and Java version seem to be outdated:

- GeneticAlgorithm.java line 26 mentions `ORLIB_FILE` while in the `weing`-format test
- cmultiknapsack.cpp lines 1508 and 1515 are identical (but don't seem to really add anything anyway)

Furthermore, the Java version also sets `Constants.DIFF_TOLERANCE = 5000;` (line 31) only for the orlib-files, but I can't find that same change being made in the C++ version. Is this intended behavior?
