### 2019-10-13 99bd09f
Fix number format exception
Fix number format exception
