+++ java/mknap1-1.txt
@@ -1,0 +1,14 @@
+ 6 10 3800
+ 100 600 1200 2400 500 2000
+ 8 12 13 64 22 41
+ 8 12 13 75 22 41
+ 3 6 4 18 6 4
+ 5 10 8 32 6 12
+ 5 13 8 42 6 20
+ 5 13 8 48 6 20
+ 0 0 0 0 8 0
+ 3 0 4 0 8 0
+ 3 2 4 0 8 4
+ 3 2 4 8 8 4
+ 80 96 20 36 44 48 10 18 22 24
+ 
