### 2019-10-13 shah314/gamultiknapsack#3 @shah314
@sjvrijn I think I have fixed the number format exception. The problem was that objectives and constraints had floating point numbers and the Java code was expecting integers. I think this is solved now.
