### 2019-10-13 shah314/gamultiknapsack#2 @shah314
@sjvrijn I think this is an issue with the distribution of the constraints. The code was expecting at least 5 items to be removed in the local improvement phase. But there were less than 5 items in the knapsack. I have changed the Constants.REMOVE to 1 and this fixed the issue.
