### 2019-10-13 #1756 @shah314
@sjvrijn I replied to the issues on the git repository page. I think I have fixed both issues (the number format exception and the code hanging). Please do a git pull and review and verify. Thank you for the detailed review.
