### 2019-10-14 #1756 @shah314
@sjvrijn About the Constants.DIFF_TOLERANCE, it is not really required. But the distribution of the values in weing and orlib formats are quite different. So, I noticed that a lower value of DIFF_TOLERANCE sometimes causes the calculation of lagrangian multipliers to take a very long time for orlib formatted files. But the C++ code runs fine without it. No harm in keeping it there though.
