### 2019-10-17 shah314/gamultiknapsack#2 @sjvrijn
Thanks for the update, this issue is now solved 👍
