### 2019-10-18 #1756 @arfon
@shah314 - could you remove the following line from the end of the paper as the repository is linked in the left-hand column:

```
The Java and C++ code are available here: https://github.com/shah314/gamultiknapsack.
```

Once you've made that update, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
